[FE] refactor: dashboard 관련 파일 한 폴더로 이동
아래 파일들 ./dashboard로 이동
1. DashboardPieChart.jsx
2. DashboardTextExpense.jsx
3. DashboardVisualExpense.jsx
